+++ test/conf/new/protractor.local.android.conf.js
@@ -1,6 +1,8 @@
 let config = require('../protractor.shared.conf.js').config;
 
-config.specs = ['../../new.mobile.spec.js'];
+config.specs = [
+	'../../new.mobile.spec.js'
+];
 
 config.seleniumAddress = 'http://localhost:4726/wd/hub';
 
@@ -14,13 +16,20 @@ config.capabilities = {
 	// see https://github.com/appium/appium/issues/11083
 	// That's why it is on
 	nativeWebScreenshot: true,
+	norReset: true,
 	chromeOpts: {
-		args: ['--no-first-run'],
+		args: [ '--no-first-run' ],
 	},
 };
 
 config.plugins = [{
-  path: '../../build/index.js',
+	path: '../../../build/index.js',
+	options:{
+		baselineFolder: './localBaseline',
+		debug: false,
+		formatImageName: `{tag}-{logName}-{width}x{height}`,
+		screenshotPath: '.tmp/'
+	}
 }];
 
 exports.config = config;