+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## Breaking changes
 - create a `canvasScreenshot` of an element has been removed
 - `saveFullPageScreens` has been renamed to `saveFullPageScreen`
-
+- `nativeWebScreenshot` has been removed from the options. This can only be added/determined from the capabilities
 
 ## NEW
 - `resizeDimensions` for `saveElement` and `checkElement` will now accept an object in pixels, see below
@@ -18,6 +18,8 @@
   }
 }
 ```
+
+- `savePerInstance` for the main options. This will save each screenshot in a subfolder which is named after the desktop `browserName` or if mobile the `deviceName` from the capabilities
 
 
 ## Deprecated